Output types added: Log/Warning/Error/Network
