+++ Assets/DeveloperConsole/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da8d67347d2653941b2a90e3173043ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
