+++ Assets/DeveloperConsole/Misc.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 47c36341c7a287b4d9c16f029a111928
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
