+++ Assets/DeveloperConsole/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 62a99ff0dd8bb9b41b07fe55ef207cf0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
