+++ Assets/Scenes/SampleScene.unity
@@ -292,8 +292,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7e306f2b190492640810c1a9383a9768, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  active: 1
   scrollPosition: {x: 0, y: 0}
+  lineSpacing: 20
+  inputLimit: 512
+  tex_thrashcan: {fileID: 2800000, guid: 48ee80cded7873a42bc3ddc483f59086, type: 3}
   skin: {fileID: 11400000, guid: 8b732f0f5154e144f8f5d0b3fd9cddf9, type: 2}
+  input: Command here
 --- !u!4 &1301172865
 Transform:
   m_ObjectHideFlags: 0
